refactor com.javahelp in app module to com.javahelp.frontend
