+++ app/src/androidTest/java/com/javahelp/frontend/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.javahelp;
+package com.javahelp.frontend;
 
 import android.content.Context;
