+++ app/src/main/java/com/javahelp/frontend/MainActivity.java
@@ -1,8 +1,10 @@
-package com.javahelp;
+package com.javahelp.frontend;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.javahelp.R;
 
 public class MainActivity extends AppCompatActivity {
