+++ app/src/test/java/com/javahelp/frontend/ExampleModelUnitTest.java
@@ -1,4 +1,4 @@
-package com.javahelp;
+package com.javahelp.frontend;
 
 import static org.junit.Assert.assertEquals;
