+++ app/src/test/java/com/javahelp/frontend/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.javahelp;
+package com.javahelp.frontend;
 
 import org.junit.Test;
